mlops-multi-account-cdk/mlops-infra/requirements.txt
aws-cdk-lib>=2.12.0
aws_cdk.aws_lambda_python_alpha
constructs>=10.0.0,<11.0.0
yamldataclassconfig
